refactor(agent): extract registry store from profile into dedicated module
